requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
sentry-sdk>=1.40,<2.0
django-prometheus>=2.3,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
requests>=2.31,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
